Update checksum digit finding
So, thanks to Sam Landfried, I got the homework done in v2 of this. I
kept v1, because I learned a lot in it. But it could technically
generate a checksum from an account number, and append it to the number.

Alternatively, it could test for a valid checksum digit.

These functionalities are not well called-out right now, but most of the
work is done. Since it's not homework, I'm done working on this.

:)